Major folder structure refactoring
