+++ src/RetroDev.OpenUI/Core/Windowing/ThreadDispatcher.cs
@@ -1,0 +1,9 @@
+﻿namespace RetroDev.OpenUI.Core.Windowing;
+
+/// <summary>
+/// Estabilish what operation is done in what thread.
+/// </summary>
+public class ThreadDispatcher
+{
+
+}
