+++ src/RetroDev.OpenUI/UI/Themes/ThemeParser.cs
@@ -1,8 +1,8 @@
-﻿using RetroDev.OpenUI.Exceptions;
-using System.Xml.Linq;
+﻿using System.Xml.Linq;
 using System.Xml;
 using System.Diagnostics.Contracts;
 using RetroDev.OpenUI.Core.Graphics;
+using RetroDev.OpenUI.UI.Themes.Exceptions;
 
 namespace RetroDev.OpenUI.UI.Themes;
