+++ src/RetroDev.OpenUI/UIDefinition/EAMLBinder.cs
@@ -1,8 +1,8 @@
 ﻿using System.Reflection;
 using System.Text.RegularExpressions;
 using RetroDev.OpenUI.Components.Base;
-using RetroDev.OpenUI.Exceptions;
 using RetroDev.OpenUI.UI.Properties;
+using RetroDev.OpenUI.UIDefinition.Exceptions;
 
 // TODO: bindings
 // @dictionaryKey.property
