+++ src/RetroDev.OpenUI/UIDefinition/IEAMLBinder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Reflection;
-using RetroDev.OpenUI.Exceptions;
 using RetroDev.OpenUI.UI.Properties;
 using RetroDev.OpenUI.Components.Base;
+using RetroDev.OpenUI.UIDefinition.Exceptions;
 
 namespace RetroDev.OpenUI.UIDefinition;
